+++ lib/reflex_tools_flooding.py
@@ -9,8 +9,8 @@ Author(s):     Mauro Arcorace (mauro.arcorace@cimafoundation.org)
                Giulia Bruno (giulia.bruno@cimafoundation.org)
                Alessia Matanò
                Andrea Libertino (andrea.libertino@cimafoundation.org)
-Date:          '20220913'
-Version:       '2.0.2'
+Date:          '20230330'
+Version:       '2.1.0'
 """
 from scipy.optimize import minimize
 from numpy.random import rand
@@ -46,15 +46,18 @@ def optimise_volume(stream_id, optimise_setting, d):
     stream_row = streams_gdf[streams_gdf["stream"] == stream_id]
 
     hand_type = "ext_no_ce"
+    c_exp = "not active"
 
     if optimise_setting["coastal_expansion_active"]:
         if optimise_setting["coastal_expansion_manual"]:
             if stream_row.manual_ce.values[0] == 1:
                 hand_type = "ext_ce"
+                c_exp = "active"
         elif stream_row.next_strea.values[0] <= 0 or stream_row.gradient.values[0] <= optimise_setting["coastal_expansion_gradient_limit"]:
             hand_type = "ext_ce"
+            c_exp = "active"
 
-
+    print(" ---> Coastal expansion mode: " + c_exp)
     hand_filename = optimise_setting["rst_hand"].format(mask_type=hand_type, stream_id=str(stream_id))
     hand_map = rxr.open_rasterio(hand_filename) / 100
 
@@ -72,10 +75,19 @@ def optimise_volume(stream_id, optimise_setting, d):
 
     print(" ---> Compute transit time...")
 
+    try:
+        roughness_coeff = stream_row["roughness"].values
+        if not roughness_coeff > 0:
+            raise ValueError
+        print(" ---> Stream roughness value provided!")
+    except:
+        roughness_coeff = optimise_setting["roughness_coeff"]
+        print(" ---> Use roughness value from config!")
+
     # compute transit time
     avg_depth = np.nanmean(flood_extent)
 
-    v_i = (1 / optimise_setting["roughness_coeff"]) * (avg_depth ** (2 / 3)) * (stream_row["gradient"].values  ** (0.5))
+    v_i = (1 / roughness_coeff) * (avg_depth ** (2 / 3)) * (stream_row["gradient"].values  ** (0.5))
     if v_i > 0:
         t_ti = int(stream_row["str_len_km"] * (10 ** 3)  / v_i)
     else: